Add input system, controllable camera, small fixes
Adds an input system that uses GLFW callbacks to capture key presses and
mouse movement. It is built using a polling system, using a double
buffering style system to ensure that the input state is consistent for
the length of a frame. It also detects when the mouse leaves the window
to prevent sudden jumps when it reenters.

The application polls the input system to determine how to move and
rotate the camera.

Since input is now captured, the Dear ImGui display can now be toggled.
If the ImGui display is hidden, the mouse cursor will be disabled so
that the mouse will not exit the window, allowing for easier control.

TODO:
- Fix camera rotation so mouse movement cannot cause roll.
- Add key rebinding and loading
- Add ability for input system to get state change on keys (so users can
query if the key has just been pressed, rather than implementing that
themselves)